deploy: partially implement deploy commands
This includes:

 - project-deploy (partially implemented, see note below)
 - project-deploy-list

These commands just leverages the commands that come with tsuru. This is
specially good for the deploy case, where there's a lot of logic
involved in uploading files to the tsuru API.

At the moment, project-deploy doesn't validate the specified env (i.e.
to ensure that the user is deploying to the first project in the
"pipeline" - usually dev). It also doesn't support version based support
it. These stuff are coming tomorrow :)

This is also missing the integration tests. I took a different approach
in the non-integration tests: I have a global variable that I replace in
the tests. This make the tests simpler, and we can trust the upcoming
integration tests.